Moved photo files to a more relevant folder
